introduce IMediaProcessor
