+++ src/alcaeus.ts
@@ -1,41 +1,38 @@
-import {Core, JsonLd} from './Constants';
 import * as FetchUtil from './FetchUtil';
-import * as GraphProcessor from './GraphProcessor';
 import * as HydraResponse from './HydraResponse';
-import {IApiDocumentation, IHydraClient, IHydraResponse, IOperation, IRootSelector} from './interfaces';
-import {forOwn} from './LodashUtil';
-import {ResourceFactory as ResourceFactoryCtor} from './ResourceFactory';
+import {
+    IApiDocumentation, IHydraClient, IHydraResponse, IMediaTypeProcessor, IOperation, IResourceFactory,
+    IRootSelector,
+} from './interfaces';
 import {IResponseWrapper} from './ResponseWrapper';
-
-function isRdfFormatWeCanHandle(mediaType: string): boolean {
-    return !!GraphProcessor.parserFactory.create(null).find(mediaType);
-}
 
 const getHydraResponse = async (
     alcaeus: IHydraClient,
     response: IResponseWrapper,
     uri: string,
-    apiDocumentation?: IApiDocumentation,
-    typeOverrides = {}): Promise<IHydraResponse> => {
-    if (isRdfFormatWeCanHandle(response.mediaType)) {
-        const processedGraph = await GraphProcessor.parseAndNormalizeGraph(
-            await response.xhr.text(),
-            uri,
-            response.mediaType);
+    apiDocumentation?: IApiDocumentation): Promise<IHydraResponse> => {
 
-        return processResources(alcaeus, uri, response, processedGraph, apiDocumentation, typeOverrides);
+    const suitableProcessor = Object.values(alcaeus.mediaTypeProcessors)
+        .find((processor) => processor.canProcess(response.mediaType));
+
+    if (suitableProcessor) {
+        return await suitableProcessor.process(uri, response, apiDocumentation);
     }
 
     return Promise.resolve(HydraResponse.create(uri, response, null, null));
 };
 
 export class Alcaeus implements IHydraClient {
-    public resourceFactory = new ResourceFactoryCtor();
+    public resourceFactory: IResourceFactory;
 
     public rootSelectors: IRootSelector[];
 
-    constructor(rootSelectors) {
+    public mediaTypeProcessors: { [name: string]: IMediaTypeProcessor };
+
+    constructor(resourceFactory: IResourceFactory, rootSelectors: IRootSelector[]) {
         this.rootSelectors = rootSelectors;
+        this.resourceFactory = resourceFactory;
+        this.mediaTypeProcessors = {};
     }
 
     public async loadResource(uri: string): Promise<IHydraResponse> {
@@ -48,10 +45,7 @@ export class Alcaeus implements IHydraClient {
     public async loadDocumentation(uri: string): Promise<IApiDocumentation> {
         try {
             const response = await FetchUtil.fetchResource(uri);
-            const typeOverrides = {};
-            typeOverrides[uri] = Core.Vocab('ApiDocumentation');
-
-            const representation = await getHydraResponse(this, response, uri, null, typeOverrides);
+            const representation = await getHydraResponse(this, response, uri, null);
             return representation.root as any as IApiDocumentation;
         } catch (e) {
             return null;
@@ -64,70 +58,4 @@ export class Alcaeus implements IHydraClient {
 
         return await getHydraResponse(this, response, uri, apiDocumentation);
     }
-}
-
-function processResources(
-    alcaeus: IHydraClient,
-    uri,
-    response,
-    resources,
-    apiDocumentation,
-    typeOverrides = {}): IHydraResponse {
-    const resourcified = {};
-    resources.forEach((res) => {
-        try {
-            res[JsonLd.Id] = new URL(res[JsonLd.Id]).href;
-        } catch (e) {}
-
-        resourcified[res[JsonLd.Id]] = res;
-    });
-
-    resources.reduceRight((acc: object, val) => {
-        const id = val[JsonLd.Id];
-        acc[id] = alcaeus.resourceFactory.createResource(alcaeus, val, apiDocumentation, acc, typeOverrides[id]);
-        return acc;
-    }, resourcified);
-
-    forOwn(resourcified, (resource) => resourcify(alcaeus, resource, resourcified, apiDocumentation, typeOverrides));
-
-    return HydraResponse.create(uri, response, resourcified, alcaeus.rootSelectors);
-}
-
-function resourcify(alcaeus: IHydraClient, obj, resourcified: object, apiDoc: IApiDocumentation, typeOverrides) {
-    if ((typeof obj === 'object') === false) {
-        return obj;
-    }
-
-    if (obj[JsonLd.Value]) {
-        return obj[JsonLd.Value];
-    }
-
-    const selfId = obj[JsonLd.Id];
-
-    if (!selfId) {
-        return obj;
-    }
-
-    let resource = resourcified[selfId];
-    if (!resource || typeof resource._processed === 'undefined') {
-        const id = obj[JsonLd.Id];
-        resource = alcaeus.resourceFactory.createResource(alcaeus, obj, apiDoc, resourcified, id);
-        resourcified[selfId] = resource;
-    }
-
-    if (resource._processed === true) {
-        return resource;
-    }
-
-    resource._processed = true;
-    forOwn(resource, (value, key) => {
-        if (Array.isArray(value)) {
-            resource[key] = value.map((el) => resourcify(alcaeus, el, resourcified, apiDoc, typeOverrides));
-            return;
-        }
-
-        resource[key] = resourcify(alcaeus, value, resourcified, apiDoc, typeOverrides);
-    });
-
-    return resource;
 }
