+++ tsconfig.json
@@ -13,6 +13,7 @@
     "suppressImplicitAnyIndexErrors": false,
     "lib": [
       "es2017",
+      "es2015",
       "dom"
     ]
   },
